Move EnemyRenderer into enemies/ directory.
